+++ src/fintrist_lib/learn.py
@@ -166,18 +166,18 @@ class Trainer():
                 self.optimizer.step()
 
                 running_loss += loss.item()
-            else:
-                loss_metric = running_loss/len(self.traindata)
-                current_lr = self.optimizer.param_groups[0]['lr']
-                metrics = {
-                    'epoch': self.epoch,
-                    'loss': loss_metric,
-                    'lr': current_lr,
-                }
-                self.performance.append(metrics, ignore_index=True)
-                print("Epoch: {epoch}, Training loss: {loss}, LR: {lr}".format(**metrics))
-                self.scheduler.step()
-                # self.scheduler.step(running_loss/len(self.traindata))
+
+            loss_metric = running_loss/len(self.traindata)
+            current_lr = self.optimizer.param_groups[0]['lr']
+            metrics = {
+                'epoch': self.epoch,
+                'loss': loss_metric,
+                'lr': current_lr,
+            }
+            self.performance.append(metrics, ignore_index=True)
+            print("Epoch: {epoch}, Training loss: {loss}, LR: {lr}".format(**metrics))
+            self.scheduler.step()
+            # self.scheduler.step(running_loss/len(self.traindata))
         ## Store training state
         self.save_state()