+++ CHANGELOG.rst
@@ -12,7 +12,8 @@ Added
 
 Fixed
 -----
-* Fix populating NN performance df. 
+* Fix populating NN performance df.
+* Update Pytorch dependency to >=1.6 to fix random_split seeding.
 
 Changed
 -------