+++ CHANGELOG.rst
@@ -11,7 +11,8 @@ Added
 * In learn.py, DFData does train/test split.
 * NN Trainer does validation on holdout test data set.
 * Can now transfer Study data to an archive dict.
-* Method for deleting orphaned GridFS file data.
+* Add method for deleting orphaned GridFS file data.
+* Add NN prediction function.
 
 Fixed
 -----